// project/src/models/enums/HideoutEventActions.ts
export enum HideoutEventActions
    {
    HIDEOUT_UPGRADE = "HideoutUpgrade",
    HIDEOUT_UPGRADE_COMPLETE = "HideoutUpgradeComplete",
    HIDEOUT_PUT_ITEMS_IN_AREA_SLOTS = "HideoutPutItemsInAreaSlots",
    HIDEOUT_TAKE_ITEMS_FROM_AREA_SLOTS = "HideoutTakeItemsFromAreaSlots",
    HIDEOUT_TOGGLE_AREA = "HideoutToggleArea",
    HIDEOUT_SINGLE_PRODUCTION_START = "HideoutSingleProductionStart",
    HIDEOUT_SCAV_CASE_PRODUCTION_START = "HideoutScavCaseProductionStart",
    HIDEOUT_CONTINUOUS_PRODUCTION_START = "HideoutContinuousProductionStart",
    HIDEOUT_TAKE_PRODUCTION = "HideoutTakeProduction",
    HIDEOUT_RECORD_SHOOTING_RANGE_POINTS = "RecordShootingRangePoints",
    HIDEOUT_IMPROVE_AREA = "HideoutImproveArea"
}